perf: Various performance improvements in the algorithm package
We now use an int to store the best schedule as the block that accesses this variable is synchronised.

In Schedule, we don't perform our own deep copies anymore.

In Scheduler, we now check if the current schedule contains a key when we want to check if a node is scheduled.